Convert Faruq plugin from kts to kt
